kanban-api/requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
asttokens==2.4.0
backports.zoneinfo==0.2.1
cffi==1.16.0
colorama==0.4.6
cryptography==41.0.4
dj-rest-auth==5.0.1
Django==4.2.5
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
executing==2.0.0
factory-boy==3.3.0
Faker==19.6.2
icecream==2.1.3
pycparser==2.21
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
